libgles-omap3: fix RPM compatibility by faking hardcoded deps
When RPM is used as the packaging type, rootfs will fail because RPM, by
design, check for file level dependency. Here one of the shared objects
(libpvrEWS_REMWSEGL.so) has dependency on ld-linux.so.3 and libews.so.

It is unclear why meta-ti requires ld-linux.so.3, which is a dynamic
loader for soft-float arch. Hence I fake the existence here.

Signed-off-by: Adam YH Lee <adam.yh.lee@gmail.com>
Signed-off-by: Denys Dmytriyenko <denys@ti.com>