linux-ti-staging: Work around to avoid race condition with external module
Currently there is an issue with compiling external kernel module
which depends on symbols from another kernel module which is part of the
list of modules configured to be built as modules in the default config.

This work around allows to avoid the issue by copying the module
symbols to the kernel directory after making the kernel modules build.

This issue is already fixed in oe-core: master branch as of date
 - this is just work around before updating to latest oe-core

Signed-off-by: Sam Nelson <sam.nelson@ti.com>
Signed-off-by: Denys Dmytriyenko <denys@ti.com>