pru-icss: add new device support and bump SRCREV
The PRU Software Support Package now includes examples for
the k2g device that need to be built and then placed into
the file system.

The Linux driver now supports loading firmwares into both
PRU-ICSS subsystems on the am437x device so the recipe was
also updated to build and install those firmwares.

Non-backward compatible changes were made to the inner
workings of rpmsg that required a major revision bump. The
package revision was bumped to 5.0.0 and this commit moves
the SRCREV to the v5.0.0 tag.

Signed-off-by: Jason Reeder <jreeder@ti.com>
Signed-off-by: Denys Dmytriyenko <denys@ti.com>