pru-icss: bump SRCREV, PV, and PR due to updates
Due to bug fixes and updates, the package version number
is being incremented to 5.0.1. The commit ID is also being
updated to reflect these changes.

Line ending corrections to the manifest file also caused
a new checksum value to be necessary. This has been updated
in this commit as well.

Signed-off-by: Jason Reeder <jreeder@ti.com>
Signed-off-by: Denys Dmytriyenko <denys@ti.com>