ti-sgx-ddk-um: add support for importing DMABUF as EGLImage
This patch adds support for importing DMABUFs as EGLImages.
This work is based on the Spec defined in
https://www.khronos.org/registry/egl/extensions/EXT/EGL_EXT_image_dma_buf_import.txt

The feature currently supports YUYV and NV12 formats
In addition, the following are supported:
* YUV full range and narrow range
* YUV color space BT709 and BT601
* YUV color space BT2020 is not supported
* Chroma siting parameters are ignored

For YUYV buffers, user needs to pass the width, height, and a 3-tuple
consisting of the dmabuf fd, pitch, and offset=0.

For YUV buffers, user needs to pass the width, height and two 3-tuples
consisting of the dmabuf fds, pitches, and offsets=0.
Please note that multiple dmabuf fds are not supported, and therefore,
for NV12,
* fd[0] must be equal to fd[1] and
* pitch[0] must be equal to pitch[1]

Signed-off-by: Anand Balagopalakrishnan <anandb@ti.com>
Signed-off-by: Denys Dmytriyenko <denys@ti.com>