uim: Fix build breaks when toolchain is used from sstate
CC is being redefined here but its not considering
the essential options to specify the sysroot as a result
this breaks on systems tied to sstate servers where the toolchain
might have had a different sysroot when built, in here
it fails to find the standard headers due to missing sysroot
options.

Deleting the extra oeconf works fine.

Signed-off-by: Khem Raj <raj.khem@gmail.com>
Signed-off-by: Denys Dmytriyenko <denys@ti.com>