ti-sgx-ddk-um: fix a bug at the drmModeAddFB() call at null_drm_ws.c
Both parameter depth and bpp are set to uiPixelWidthInBits at the current
implementation, however, the depth may not be the same as bpp for certain
pixel format such as XRGB8888. This bug was exposed by the recent DRM
driver change introducing strict error checking. To fix this problem,
add utility function uiPixelDepthInBits to derive and use depth based on
the pixel format.

Signed-off-by: Eric Ruei <e-ruei1@ti.com>
Signed-off-by: Denys Dmytriyenko <denys@ti.com>