recipes-kernel:cmem.dtsi:K2G: Update to remove redefinition of mpm areas
Cleaning up redefinition of mpm areas and moving to use what is defined
in the base device tree files as is.
This effectively changes to a new memory partition.
Note that this change needs a corresponding change in the dts files in
the kernel tree to adjust the mpm memory area to the new memory
partition.

The CMEM areas here are adjusted such that there is no overlap with the
mpm areas defined in the dts files in the kernel tree.
Note also the address ranges are arranged differently for K2G compared
to  the other Keystone platforms.
(This is to accomodate common memory map for multiple k2g platforms)

Signed-off-by: Sam Nelson <sam.nelson@ti.com>
Signed-off-by: Denys Dmytriyenko <denys@ti.com>