amx3-cm3: Fix building with poky-lsb
One of the features of poky-lsb is that it brings in the
security_flags.inc file that sets a number of security-related compile
time flags.  This includes changing some of the default GCC build
options.  However, for building the amx3-cm3 firmware we want to ensure
that none of this enabled as it's not useful in this particular
environment.  We must pass in ${SECURITY_NOPIE_CFLAGS} to disable PIE.
When security_flags.inc is not present this is an empty string so is
safe to do at all times.

Signed-off-by: Tom Rini <trini@konsulko.com>
Signed-off-by: Denys Dmytriyenko <denys@ti.com>