ti-sgx-ddk-um: libgbm is now part of this project
The GBM library is tightly coupled with the system graphics systems,
other hardware GPU vendors (including our Rogue DDK) and Mesa all deploy
a version of the library as part of their GPU driver. We SGX does the
same. This should ease integration for us as we now don't rely on a
separate project for this library and our deployment strategy can match
across Mesa/RGX recipes, allowing for better re-use of common components.

Signed-off-by: Andrew F. Davis <afd@ti.com>
Signed-off-by: Denys Dmytriyenko <denys@ti.com>