ti-sgx-ddk-um: Add expat as runtime dependency
Although not needed for the IMG default EGL implementation libexpat
is needed when using the Mesa EGL version.

While here s/libudev/udev as 'libudev' is package shipped as part of the
eudev project, but we use systemd, so we actually depend on 'udev'.

Move the PACKAGE_ARCH and PR defines to a more natural spot that also
lines up with the RGX version of recipe for easier diffs.

One thing we could have done is remove the DEPENDS line. We do not have
any build dependencies as these are pre-made library binaries, but I use
the package rootfs generated by bitbake to build the binaries locally,
so leave it for now.

Signed-off-by: Andrew F. Davis <afd@ti.com>
Signed-off-by: Denys Dmytriyenko <denys@ti.com>