trusted-firmware-a: Update to v2.4 merge
Update to v2.4 tag merge from upstream to keep TF-A in sync.

For major changes in trusted-firmware-a, See:
https://trustedfirmware-a.readthedocs.io/en/latest/change-log.html#version-2-4

TI specific fixups include:
 - stub for shutdown (CPUs enter wfi)
 - uart console is now configurable.

While at it, update to include PV_k3 as 2.4. We use PV_k3 instead of
PV since this allows meta-ti to not affect other platforms esp in a
multi-BSP configuration.

We also add LIC_FILES_CHKSUM_k3 to make sure that we don't let any
re-licensing in the project pass through unnoticed as that has to
be carefully examined for making sure we have all the right legal
compliance - There has been no change compared to previous versions,
but we introduce this as a protection in case there is a change in the
future.

Reviewed-by: Denys Dmytriyenko <denis@denix.org>
Suggested-by: Dan Murphy <dmurphy@ti.com>
Signed-off-by: Nishanth Menon <nm@ti.com>
Signed-off-by: Dan Murphy <dmurphy@ti.com>