linux-ti-staging, u-boot-ti-staging: lower default preference
Latest versions of Linux kernel (5.10) and U-boot (2021.01) are not yet ready
for prime time in meta-ti and some features are still being migrated and causing
issues. The recipes are meant for development and not yet for production.

Normally selection of the version is done with PREFERRED_VERSION by the Distro
(e.g. Arago), but if not specified by a Distro (Poky, distroless, etc.), the
default is to use the latest version. Hence need to change the default to not
automatically pick the latest version that is not ready yet.

This can be removed/reverted once those versions are ready to be picked by
default.

Signed-off-by: Denys Dmytriyenko <denis@denix.org>
Signed-off-by: Dan Murphy <dmurphy@ti.com>