linux-ti-mainline: Set the branch for the ti-upstream-tools repo
ti-upstream-tools[1] repository carries the overlays and defconfig
fragments that we don't wish to overload upstream kernel community with.
The reason for that being these are constantly on churn as our test
coverage increases etc. Currently we reuse the "BRANCH" variable to also
matchup with kernel and ti-upstream-tools repositories as well.

This can tend to be a little painful if we decide to start testing
various kernel configurations - RT etc.. So, lets just split out the
TOOLS_BRANCH separated from the branch we use for kernel. Even though
the branch remains the exact same at this point.

[1] https://git.ti.com/cgit/ti-linux-kernel/ti-upstream-tools/

Signed-off-by: Dan Murphy <dmurphy@ti.com>
Signed-off-by: Nishanth Menon <nm@ti.com>