conf/machine: k3: Add multi-certificate boot image support
Lets keep things consistent by providing two possibilities for platforms
to pick from - legacy boot and multi-certificate boot.

In legacy boot, the base bootloader and system firmware are
maintained as separate binaries (tiboot3.bin and sysfw.itb).

In multi-certificate boot that newer K3 devices support, ROM is smarter
and can handle multiple x509 certificate based images: so we can strip
out the sysfw.itb and integrate it as part of tiboot3.bin itself. This
improves authentication and overall system boot times since we are now
able to boot both the system controller and the boot processor in
parallel.

We do have a scheme currently to identify the images necessary for boot
etc, but things are handled on a platform conf file basis. We can
improve that by introducing the pattern at the top level include and use
the relevant pattern in platforms as needed.

Signed-off-by: Nishanth Menon <nm@ti.com>
Reviewed-by: Denys Dmytriyenko <denys@konsulko.com>