machine: k3r5: Make combined R5 boot the default
Going forward, the combined tiboot3.bin is going to be the standard. Only
AM65x and J721e will still use the split boot. Set the output of U-Boot
to be u-boot-spl.bin by default and override only for the two split boot
platforms. This makes it easier to add new platforms and simplifies
deployment logic.

Signed-off-by: Andrew Davis <afd@ti.com>
Signed-off-by: Ryan Eatmon <reatmon@ti.com>