u-boot-ti-staging: Update to the latest ti-u-boot-2021.01
Update to the latest commit of ti-u-boot-2021.01.

As part of this U-Boot update the non-HS AM64x-evm defconfig was removed.
This is okay as building this defconfig without defining TI_SECURE_DEV_PKG
will still result in usable binaries for the non-HS devices. Use the
HS defconfig for both the HS and non-HS machines.

Signed-off-by: Andrew Davis <afd@ti.com>
Signed-off-by: Ryan Eatmon <reatmon@ti.com>