optee-os: Lower log level for AM62x
During resume from low power mode suspend OP-TEE prints some messages
on the UART. It seems this UART is not powered at this point in the
sequence breaking suspend/resume. We should track down the exact
prints and quiet them. Until we get that fix upstream, lets lower
the log level for this platform to unblock the LPM work.

Signed-off-by: Andrew Davis <afd@ti.com>
Acked-by: Anand Gadiyar <gadiyar@ti.com>
Signed-off-by: Ryan Eatmon <reatmon@ti.com>