ti-sci-fw: Make combined R5 boot the default
Going forward, the combined tiboot3.bin is going to be the standard. Only
AM65x and J721e will still use the split boot.

Set DEPENDS and EXTRA_OEMAKE for combined boot and override only for the
two split boot platforms. This makes it easier to add new platforms and
simplifies deployment logic.

Signed-off-by: Andrew Davis <afd@ti.com>
Signed-off-by: Ryan Eatmon <reatmon@ti.com>
Signed-off-by: Denys Dmytriyenko <denys@konsulko.com>
Signed-off-by: Ryan Eatmon <reatmon@ti.com>