j721s2 and j784s4: Correct serial console device
Serial console in J721S2/J784S4 EVM config was set to ttyS10.  It should be ttyS2.
This caused sysv based systems (eg, tiny) to apparently hang because of no login prompt.
This didn't show up as a problem on systemd systems (eg, default) because systemd will
automatically setup a serial tty with the console option passed in on the kernel command line
(using console=...), which U-Boot correctly passes in as ttyS2 for J721S2/J784S4.

Signed-off-by: Jonathan Humphreys <j-humphreys@ti.com>
Signed-off-by: Ryan Eatmon <reatmon@ti.com>
Signed-off-by: Denys Dmytriyenko <denys@konsulko.com>
Signed-off-by: Ryan Eatmon <reatmon@ti.com>