ti-sci-fw: make dependency on meta-ti-extras soft
Replace "require" with "include" to load ti-paths.inc and not break
parsing when meta-ti-extras is not in the BBLAYERS list.

As ti-paths.inc is only needed to define TI_K3_SECDEV_INSTALL_DIR
for ti-k3-secdev when TI_SECURE_DEV_PKG_K3 is not set and building
for K3 HS platforms. Since ti-k3-secdev resides in meta-ti-extras,
the soft dependency will still work fine and give a build error
"Nothing PROVIDES 'ti-k3-secdev-native'" when it is needed.

Also, since TI_K3_SECDEV_INSTALL_DIR can now be undefined, use
d.getVar() to access it ensuring it doesn't break the condition.

Reported-by: Daniel Díaz <mrchapp@gmail.com>
Signed-off-by: Denys Dmytriyenko <denys@konsulko.com>
Signed-off-by: Ryan Eatmon <reatmon@ti.com>