ti-rtos-firmware: Use ti-k3-secdev if TI_SECURE_DEV_PKG_K3 not defined
Use ti-k3-secdev native package to provide secure-binary-image.sh script
when it is not passed on as an environment variable. This fixes and issue
with AM64xx which is always HS but should not require TI_SECURE_DEV_PKG
when building for HS-FS platforms.

This is backported from a patch for kirkstone. [1]

[1] https://lists.yoctoproject.org/g/meta-ti/topic/master_kirkstone_patch_1_8/96819771

Signed-off-by: Ryan Eatmon <reatmon@ti.com>