conf: machine: k3: Define RD, DTB and DTBO load address for FIT image
K3 platforms include the FIT image-based kernel by default. The bootloader
loads the kernel and DTB from the FIT image. The DTB/DTBOs are added to the
FIT image but the load address is not included in the image section of
each DTB/DTBO. This load address is determined by the LOADADDRESS macros
in the kernel-fitmage class.

Hence, define the RD/DTB/DTBOs LOADADDRESS and OFFSET macros in the K3 common
include, and remove those macros defined in the K3 board specific config files.

Signed-off-by: Sinthu Raja <sinthu.raja@ti.com>
Reviewed-by: Andrew Davis <afd@ti.com>
Signed-off-by: Ryan Eatmon <reatmon@ti.com>