ti-rtos-firmware: Use a table instead of a list of copy commands
Using a table makes it easier to see what firmware is included with
each machine build. It also helps prevent errors here.

Since the same functions run across machines we now sign all the firmware
without needing to add a new signing function for each new machine.

Another fix is that we install the DM firmware to the deploy directory
for all machines that use need it. This was added for J7x machines
but missed for AM62x.

Signed-off-by: Andrew Davis <afd@ti.com>
Signed-off-by: Ryan Eatmon <reatmon@ti.com>