u-boot-ti-staging: Fix u-boot configs to match oe-core fitImage
After moving to the oe-core kernel-fitimage.bbclass we found that the
dtb names we were searching for did not match.  This was due to the
KERNEL_DEVICETREE entries including the vendor subdir in the section
names of the fitImage file.  This patch updates support in 2021 so that
we can temporarily get to a stable boot point before starting our LTS
migration and changing a lot of things in the recipes.

Signed-off-by: Ryan Eatmon <reatmon@ti.com>