kernel-fitimage: Add signing of fitImage entries for 5.10 kernel
While we are mirgrating to kernel v6.1 and u-boot v2023.04, we want to
keep kernel v5.10 and u-boot v2021.01 "working" for anyone looking at
kirkstone.  One of the items we are looking at changing is the signing
of entries in the fitImage.

To try and acheive a limited parity with dunfell while we work on the
migration, this commit creates a class that implements the logic that was in
dunfell and applies that class to the 5.10 kernel.

This is a temporary patch that will be remvoed when we remove the 5.10
recipe down the road.  This logic will not apply to the 6.1 kernel.

Signed-off-by: Ryan Eatmon <reatmon@ti.com>