ti-rtos-firmware: Use the same directory structure as linux-firmware
Currently we move this firmware around and put it in a directories named
pdk-ipc/ and ethfw/. This does not match what is in linux-firmware which
is what projects should expect /lib/firmware to match. This causes issues
for projects that want to work both on meta-ti and also on any other distro.

Switch to shipping the same directory structure as linux-firmware here.

Signed-off-by: Andrew Davis <afd@ti.com>
Signed-off-by: Ryan Eatmon <reatmon@ti.com>