conf: machine: Add extra tiboot3.bin version that matches default
When creating the boot partition we copy all of the images for the
different board types (gp, hs-fs, hs-se), but one of those we consider
to be a default.  That default is not copied with the longer name, but
rather with the tiboot3.bin name.

The issue comes when someone wants to boot with a different bin file.
They need to copy the approriate longer named .bin file to tiboot3.bin.
But then the original name of what tiboot3.bin is lost unless the use
backs it up.

This patch just additionally copies that longer named bin file so that
anyone who wants to look at the boot partiion and decide to boot for a
different board can know which file to copy to tiboot3.bin and does not
need to worry about making a backup.

Signed-off-by: Ryan Eatmon <reatmon@ti.com>