ti-ipc: Stop looking for rpmsg_rpc.h in the kernel
Stop looking for the rpmsg_rpc.h in the kernel.  While we are doing the
LTS migration for kirkstone, this file is not going to be in the kernel
for a bit.  This patch works in conjunction with a recipe change that
downloads the header and adds a -I to the CFLAGS to pick it up.  This
will be removed once the 6.1 kernel has been updated.

Signed-off-by: Ryan Eatmon <reatmon@ti.com>