ti-img-rogue-umlibs: drop the api specific dev packages
Well, flew a bit too close to the sun on that last patch. Apparently
there are a few hidden runtime dependencies on the API devlinks that
make things like EGL upset when they aren't present.

We may circle back to this at some point but for now we're fine shipping
the devlinks in the base API packages as these links don't step on any
other package.

Fixes: 1899f5ea (ti-img-rogue-umlibs: bump version and explicitly
package, 2023-06-01)

Signed-off-by: Randolph Sapp <rs@ti.com>
Signed-off-by: Ryan Eatmon <reatmon@ti.com>