ti-extras: Add support for extra kernel/u-boot features
There are some features that will either never be accepted into the
upstream kernel, or are not ready to be accepted.  ti-extras allows for
controlling the kernel and u-boot repos, breanches, and srcrevs to gain
access to those features via the TI_EXTRAS variable in the local.conf
file.

Initial support for TI_EXTRAS=tie-jailhouse is being added by this
patch.

Signed-off-by: Paresh Bhagat <p-bhagat@ti.com>
Signed-off-by: Ryan Eatmon <reatmon@ti.com>
Acked-by: Paresh Bhagat <p-bhagat@ti.com>
Signed-off-by: Ryan Eatmon <reatmon@ti.com>