conf: machine: k3: Use ARM64 kernel load address in FIT
The address 0x80008000 was used in pre-K3 devices. ARM64 requires the
kernel loaded to a 2MB aligned address when not using KASLR. For FIT
set the address to 0x81000000 which gives the kernel 32MB before running
into the FDT at 0x83000000.

If you get this error someday then your kernel is too large and either it
or FDT needs relocated in memory:
    ERROR: image is not a fdt - must RESET the board to recover.

Signed-off-by: Andrew Davis <afd@ti.com>
Tested-by: Praneeth Bajjuri <praneeth@ti.com>
Signed-off-by: Ryan Eatmon <reatmon@ti.com>