linux-ti-next,conf/machine: Add support for new vendored dts dir
The latest kernel has made the change to add vendor subdirectories into
the arch/arm/boot/dts directory.  This effectively breaks the
KERNEL_DEVICETREE settings in the machine configs for 32bit platforms,
so switch them over to using the PREFIX variable.  Also change the
linux-ti-next to use the PREFIX setting for KERNEL_DEVICETREE.

Signed-off-by: Ryan Eatmon <reatmon@ti.com>