mesa-pvr: add dynamic-layer append for llvmpipe
Meta-clang provides a bbappend for mesa to use clang to accelerate it's
rasterization with runtime code generation through the use of llvmpipe.
With the addition of mesa-pvr we no longer get this for free with the
blanket mesa bbappend on the devices that explicitly request mesa-pvr so
lets add a dynamic layer to append the same args meta-clang does.

Though the number of devices that would select pvr-mesa and still want
to use software rendering is small, it's not zero due to debug and
testing purposes.

Signed-off-by: Randolph Sapp <rs@ti.com>
Signed-off-by: Ryan Eatmon <reatmon@ti.com>