conf: include TI_COMMON_DEPLOY in PSEUDO_IGNORE_PATHS
Since the introduction of TI_COMMON_DEPLOY, we've been seeing frequent
Pseudo errors [1] breaking image build during development, in particular
with partial rebuilds due to the externalsrc bbclass.

The relevant part of the error log seems to be the following:

    ERROR: Task (/home/schifferm/Devel/src/ci-meta-tq-kirkstone/sources/meta
        -tq/meta-dumpling/recipes-images/images/tq-image-generic-debug.bb:do
	_rootfs) failed with exit code '1'
    Pseudo log:
    path mismatch [1 link]: ino 12215614 db '/home/schifferm/Devel/src/ci-me
        ta-tq-kirkstone/build-mba64xx/deploy-ti/images/tqma64xxl-mbax4xxl/tq
	-image-generic-debug-tqma64xxl-mbax4xxl-20231020091917.testdata.json
	' req '/home/schifferm/Devel/src/ci-meta-tq-kirkstone/build-mba64xx/
	tmp/work/tqma64xxl_mbax4xxl-tq-linux/tq-image-generic-debug/1.0-r0/r
	ootfs/usr/share/stress-ng/example-jobs/device.job;65324e26'.

My understanding is that all deploy dirs should be ignored by Pseudo, as
openembedded-core/meta/bitbake.conf includes the whole of DEPLOY_DIR in
PSEUDO_IGNORE_PATHS. With the introduction of TI_COMMON_DEPLOY,
DEPLOY_DIR_IMAGE is not covered by that setting anymore, so we add it in
meta-ti to avoid the issue.

[1] https://wiki.yoctoproject.org/wiki/Pseudo_Abort

Signed-off-by: Matthias Schiffer <matthias.schiffer@ew.tq-group.com>
Acked-by: Denys Dmytriyenko <denys@konsulko.com>
Signed-off-by: Ryan Eatmon <reatmon@ti.com>