u-boot-bb.org: add back Beagle-specific U-boot 2023.04
Majority of Beagle devices use ti-u-boot repo from u-boot-ti-staging directly.
But some latest devices may need to point to own tree with additional support
added that is not yet upstream, e.g. BeaglePlay.

Add the Beagle-specific recipe and point beagleplay to use it. Other devices
can switch to using it as well, if desired, but right now the code base is the
same for other Beagle devices.

Signed-off-by: Denys Dmytriyenko <denys@konsulko.com>
Signed-off-by: Ryan Eatmon <reatmon@ti.com>