conf: machine: move RDEPENDS and RRECOMMENDS from kernel-rdepends.inc
Instead of making kernel RDEPENDS on devicetree and different FW images with
lots of machine-override appends that are hard to modify downstream, use proper
MACHINE_ESSENTIAL_EXTRA_RDEPENDS and MACHINE_ESSENTIAL_EXTRA_RRECOMMENDS vars
that are specifically designed for this purpose.

This also makes such dependencies generic and not tied to a particular kernel
recipe including kernel-rdepends.inc with the entire machine-specific list.

Signed-off-by: Denys Dmytriyenko <denys@konsulko.com>
Acked-by: Andrew Davis <afd@ti.com>
Acked-by: Ryan Eatmon <reatmon@ti.com>
Signed-off-by: Ryan Eatmon <reatmon@ti.com>