ti-img-rogue-driver: Fix build error due to .SECONDARY/.NOTINTERMEDIATE
There is a build error under master that is not present in kirkstone
related to .SECONDARY/.NOTINTERMEDIATE being set differently between the
ti-img-rogue-driver repo and the kernel.  The long term solution to
allow this code to be built on both kirkstone/master is being
investigated.  In the meantime, this patch gets things building on
master.

Signed-off-by: Ryan Eatmon <reatmon@ti.com>