ti-img-rogue-*: version bump to 23.3.6512818
Migrate to version 23.3.6512818, this includes the stability updates
released in 23.2 as well as some additional modifications to help with
userspace application support, namely Chromium.

We're also dropping the x11 window system extensions from the Vulkan
libraries for now. Ideally, both x11 and wayland extensions will be
provided by mesa's vulkan-wsi-layer in the future, but currently x11
isn't supported and this just added extra complexity to the package in
Yocto.

Signed-off-by: Randolph Sapp <rs@ti.com>
Tested-by: Darren Etheridge <detheridge@ti.com>
Signed-off-by: Ryan Eatmon <reatmon@ti.com>