linux-ti-staging: Update KERNEL_LOCALVERSION
The use of KERNEL_LOCALVERSION for controlling the kernel version has
changed recently.  Placing the value of KERNEL_LOCALVERSION in
.scmversion was causing it to duplicate the string in the final version.

Instead, drop KERNEL_LOCALVERSION to just identify that it comes from
TI, and that it is an RT kernel.  Additionally, place the commit sha for
the kernel into the .scmversion file explicitly to be both backward and
forward compatible.

Signed-off-by: Ryan Eatmon <reatmon@ti.com>