meta-ti-bsp: Add Zephyr RTOS example firmware
Like the current TI-RTOS/MCU+ firmware, this is a prebuilt example
firmware for TI K3 remote cores. It is an alternative to the echo test
demo currently provided by "ti-rtos-firmware" recipe. In addition to
supporting rpmsg_client_sample just as the current firmware does, it
also provides:

 * A serial console with shell on the MCU UART port
 * A console over RPMSG at /dev/ttyRPMSG0
 * Example usage of the MCU GPIO/I2C/SPI from this shell
 * And many other Zephyr default enabled features..

It does all this using the Zephyr RTOS, a scalable real-time operating
system, optimized for resource constrained devices, and built with
security in mind. Zephyr is a Linux Foundation project. For more
info see here[0][1].

Zephyr firmware like this can also be built as part of the Yocto build
using meta-zephyr. Although this is a prebuilt currently, we intend
to build this from source to reduce meta-ti's dependence on firmware
blobs.

This currently works with the AM62 M4F. Support for additional SoCs
and cores is in progress.

[0] https://www.zephyrproject.org/
[1] https://docs.zephyrproject.org/latest/introduction/index.html

Signed-off-by: Andrew Davis <afd@ti.com>
Signed-off-by: Ryan Eatmon <reatmon@ti.com>