python3-yamllint: Temporarily carry this new recipe
Add recipe for yamllint.  There is an upcoming change in u-boot where
the binman tool is now configured to call yamllint to verify the configs
during compile time.

There was a previous patch a year ago from Trevor Woerner that never
made it into oe-core.  This patch is a reworking of his patch but
pointing to a newer version.

We are waiting on this patch to be accepted in oe-core.  Once it is
available on master we will remove this recipe to avoid conflict.

Signed-off-by: Ryan Eatmon <reatmon@ti.com>