u-boot-ti: deploy initial-env for the main machine
It is useful to have U-boot initial environment deployed for future
reference or to initialize environment files as part of the build.

In multiconfig setup, such as TI K3 platforms, building U-boot
multiple times for Cortex-R5 and Cortex-A53 cores results in file
conflicts when deploying under the same name. To avoid that, those
initial-env files were removed at the end of the do_deploy task.

But it is better to only remove them for k3r5 multiconfig, since
that only builds SPL and saved env doesn't make much sense, but
preserve and deploy initial-env for the main machine multiconfig.

Signed-off-by: Denys Dmytriyenko <denys@konsulko.com>
Signed-off-by: Ryan Eatmon <reatmon@ti.com>