conf: machine: add AM68-SK machine configuration
As of commit [1] there will be a separate defconfig to build u-boot
for j721s2-evm and am68-sk.

Hence, introduce new yocto machine configs for am68-sk. This is done
through a new am68.inc file as the am68-sk platform does not support GP.
So j721s2-evm.inc is copied to am68.inc and updated accordingly.

[1]: https://source.denx.de/u-boot/u-boot/-/commit/a96be9b8c05ea835d27f7c9bae03279b8bd5dcf2

Signed-off-by: Aniket Limaye <a-limaye@ti.com>
Signed-off-by: Ryan Eatmon <reatmon@ti.com>