conf: machine: Breakout IMAGE_BOOT_FILES into new variables
The IMAGE_BOOT_FILES variable has been overloaded to include additional
files that should be included in the boot partition but are not boot
loaders themselves (images, etc...).

There is also a need to generate a report in the deploy directory for
the list of actual bootloader files that were constructed as part of the
build.

Solution, break the IMAGE_BOOT_FILES into two new variables
TI_BOOTLAODERS and TI_MULTICONFIGS and default IMAGE_BOOT_FILES to
TI_BOOTLOADERS.  This way we can stick additional boot loaders into a
new variable and they will will end up in the IMAGE_BOOT_FILES list.

Signed-off-by: Ryan Eatmon <reatmon@ti.com>