classes: Create uboot-fragments.bbclass
Update our custom config fragment handling code to work with the new
UBOOT_CONFIG changes.  To better align with naming, change the name of
UBOOT_CONFIG_FRAGMENTS to just UBOOT_FRAGMENTS.  Also create a
UBOOT_FRAGMENTS_CONFIG that goes hand in hand with UBOOT_CONFIG so that
you can specify different fragments for each of the config types.

For example:

  UBOOT_CONFIG = "a b"
  UBOOT_CONFIG[a] = "config_a"
  UBOOT_CONFIG[b] = "config_b"

  UBOOT_FRAGMENTS_CONFIG[a] = ""
  UBOOT_FRAGMENTS_CONFIG[b] = "fragment1"

The logic in the class will automatically populate UBOOT_FRAGMENTS from
UBDOOT_FRAGMENTS_CONFIG.

Additionally, change the value that UBOOT_FRAGMENTS can take to be a
comma separated list of fragments to apply.

Signed-off-by: Ryan Eatmon <reatmon@ti.com>