ti-bsp: Add RT and systest support
We were missing controlling the versions for linux-ti-staging-rt and
linux-ti-staging-systest using the BSP framework.  Since the
PREFERRED_VERSION is not tied to the virtual/kernel we need two
additional settings for those alternate kernel package names.

For BSPs that do not have the -rt and -systest recipes then the two
new variable settings should effectively be ignored and cause no issues.

Signed-off-by: Ryan Eatmon <reatmon@ti.com>