linux: Rework the reproducibility patches
It was suggested to move the reproducibility patches out of
ti-kernel.inc and into the respective recipes.  Take this time to also
rename the variable being added to the SRC_URI to
REPRODUCIBILITY_PATCHES to make it clearer about what the patches are
that it lists.

This change will also make it easier for the upstream testing flow to
update the mainline recipe to the new version and update the needed
patches at the same time.

Signed-off-by: Ryan Eatmon <reatmon@ti.com>