linux: Consolidate the SRC_URI variables
It was suggested to move the reproducibility patches out of
ti-kernel.inc and into the respective recipes.

Taking this time to also rework all of the pieces that go into the SRC_URI
and create a standardization for how all of the kernel recipes are done.

Create KERNEL_GIT_URI, KERNEL_GIT_PROTOCOL, KERNEL_GIT_BRANCH,
KERNEL_DEFCONFIG, and KERNEL_REPRODUCIBILITY_PATCHES.  The default for
the above new variables is to the git.ti.com repository and how we manage
the kernel settings.  The other kernel recipes that do something other than
how TI does it can simply override the variables.

This change will make it easier to perform the upstream testing that we
do on a nightly basis.  Additionally, it will simplify the auto generation
of the patches to move the mainline (stable kernel) recipes since the
required patches will be all done in the same recipe.

Signed-off-by: Ryan Eatmon <reatmon@ti.com>