sgx: breaking things to fix other things
We got a bit more time to do a proper migration this time. A discovery
period to see what actually works and boy, it wasn't much.

This update resolves issues with:

	- Macro redefinition
	- Sparse NPOT texture allocations
	- vec4 reductions
	- Mesa image extensions
	- Buffer allocations under weston
	- Devices with both a RENDER and PRIMARY display node
	- Underprivileged users being unable to allocate buffers
	- Removed call to drm_map_legacy
	- Replace deprecated kernel calls
	- Fix some cache attributes for userspace allocations
	- About 40 other various errors and warnings

This core was never fully GLES2 conformant. In fact, if you try to run
cts on any of the older releases you'll find yourself in a hardware
recovery loop pretty quickly. We've gotten it to about 94.2% conformance
with these patches.

There are still dragons here, but they are at least a little more
manageable.

Signed-off-by: Randolph Sapp <rs@ti.com>
Signed-off-by: Ryan Eatmon <reatmon@ti.com>