conf/machine: rename beagleplay machine to beagleplay-ti
BeaglePlay is officially supported by TI "ti-6_6" BSP and is part
of the TI-SDK releases. Switching BSP to BeagleBoard.org "bb_org"
settings for the same machine is possible within the same layer.

But having different machine configs for the same machine name in
multiple layers can be confusing and prone to errors. Hence rename
the config to indicate BeaglePlay with official TI support in order
to re-use the canonical name in a separate layer.

Signed-off-by: Denys Dmytriyenko <denys@konsulko.com>